feat :p230_532 锁定限制包与版本删除；

+++ src/backend/repository/biz-repository/src/main/kotlin/com/tencent/bkrepo/repository/service/packages/impl/PackageServiceImpl.kt
@@ -39,6 +39,7 @@ import com.tencent.bkrepo.common.api.util.Preconditions
 import com.tencent.bkrepo.common.artifact.api.DefaultArtifactInfo
 import com.tencent.bkrepo.common.artifact.constant.ARTIFACT_INFO_KEY
 import com.tencent.bkrepo.common.artifact.constant.FORBID_STATUS
+import com.tencent.bkrepo.common.artifact.constant.LOCK_STATUS
 import com.tencent.bkrepo.common.artifact.constant.SCAN_STATUS
 import com.tencent.bkrepo.common.artifact.message.ArtifactMessageCode
 import com.tencent.bkrepo.common.artifact.pojo.RepositoryCategory
@@ -286,6 +287,10 @@ class PackageServiceImpl(
 
     override fun deletePackage(projectId: String, repoName: String, packageKey: String, realIpAddress: String?) {
         val tPackage = packageDao.findByKey(projectId, repoName, packageKey) ?: return
+        val tPackageVersionList = packageVersionDao.listByPackageId(tPackage.id!!)
+        if (tPackageVersionList.any { it.metadata.any { it.key == LOCK_STATUS && it.value == true } }) {
+            throw ErrorCodeException(ArtifactMessageCode.PACKAGE_LOCK, packageKey)
+        }
         packageVersionDao.deleteByPackageId(tPackage.id!!)
         packageDao.deleteByKey(projectId, repoName, packageKey)
         publishEvent(
@@ -312,6 +317,9 @@ class PackageServiceImpl(
     ) {
         val tPackage = packageDao.findByKey(projectId, repoName, packageKey) ?: return
         val tPackageVersion = packageVersionDao.findByName(tPackage.id.orEmpty(), versionName) ?: return
+        if (tPackageVersion.metadata.any { it.key == LOCK_STATUS && it.value == true } == true) {
+            throw ErrorCodeException(ArtifactMessageCode.PACKAGE_LOCK, packageKey)
+        }
         packageVersionDao.deleteByName(tPackageVersion.packageId, tPackageVersion.name)
         tPackage.versions -= 1
         if (tPackage.versions <= 0L) {
